requirements-data.txt
# Core dependencies for data loading and processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Optional: for environment variable management
python-dotenv>=1.0.0

# Optional: for technical indicators (Week 2+)
# stockstats>=0.6.0
# ta-lib>=0.4.0  # Requires C library installation

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
